requirements.txt
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
atlastk==0.13.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
fastapi==0.103.1
finnhub-python==2.4.18
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iso8601==1.1.0
nose2==0.13.0
pipreqs==0.4.13
pydantic==2.3.0
pydantic-core==2.6.3
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tortoise==0.1.1
tortoise-orm==0.20.0
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
yarg==0.1.9
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2